+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -36,5 +36,6 @@ namespace llvm{
 	class Value;
 	class Constant;
 	class ConstantInt;
+	class CallInst;
 	
 }